+++ views/tests/test-grid-1.php
@@ -1,50 +1,53 @@
+<?php
+    $PAGE_TITLE = "TEST BOOTSTRAP";
+    $pathToRootFolder = "../../";
+?>
+
 <!DOCTYPE html>
 <html lang="fr">
 
-    <?php $PAGE_TITLE = "TEST BOOTSTRAP"; ?>
-    <?php include("./views/common/head.php");?>
+<?php include($pathToRootFolder."views/common/head.php"); ?>
 
 <body>
     <!-- =================================================== -->
     <!-- ================ DEBUT HTML  ================ -->
 
-
     <h1>TEST BOOTSTRAP</h1>
 
     <div class="grid-2 grid-md-3 grid-lg-6">
-        <div class="bleu">1</div>
-        <div class="bleu">2</div>
-        <div class="bleu">3</div>
-        <div class="bleu">4</div>
-        <div class="bleu">5</div>
-        <div class="bleu">6</div>
-        <div class="bleu">7</div>
-        <div class="bleu">8</div>
-        <div class="bleu">9</div>
-        <div class="bleu">10</div>
-        <div class="bleu">11</div>
-        <div class="bleu">12</div>
-        <div class="bleu">13</div>
-        <div class="bleu">14</div>
-        <div class="bleu">15</div>
-        <div class="bleu">16</div>
-        <div class="bleu">17</div>
-        <div class="bleu">18</div>
-        <div class="bleu">19</div>
-        <div class="bleu">20</div>
-        <div class="bleu">21</div>
-        <div class="bleu">22</div>
-        <div class="bleu">23</div>
-        <div class="bleu">24</div>
-        <div class="bleu">25</div>
-        <div class="bleu">26</div>
-        <div class="bleu">27</div>
-        <div class="bleu">28</div>
-        <div class="bleu">29</div>
-        <div class="bleu">30</div>
+        <div class="bg-info m-2">1</div>
+        <div class="bg-info m-2">2</div>
+        <div class="bg-info m-2">3</div>
+        <div class="bg-info m-2">4</div>
+        <div class="bg-info m-2">5</div>
+        <div class="bg-info m-2">6</div>
+        <div class="bg-info m-2">7</div>
+        <div class="bg-info m-2">8</div>
+        <div class="bg-info m-2">9</div>
+        <div class="bg-info m-2">10</div>
+        <div class="bg-info m-2">11</div>
+        <div class="bg-info m-2">12</div>
+        <div class="bg-info m-2">13</div>
+        <div class="bg-info m-2">14</div>
+        <div class="bg-info m-2">15</div>
+        <div class="bg-info m-2">16</div>
+        <div class="bg-info m-2">17</div>
+        <div class="bg-info m-2">18</div>
+        <div class="bg-info m-2">19</div>
+        <div class="bg-info m-2">20</div>
+        <div class="bg-info m-2">21</div>
+        <div class="bg-info m-2">22</div>
+        <div class="bg-info m-2">23</div>
+        <div class="bg-info m-2">24</div>
+        <div class="bg-info m-2">25</div>
+        <div class="bg-info m-2">26</div>
+        <div class="bg-info m-2">27</div>
+        <div class="bg-info m-2">28</div>
+        <div class="bg-info m-2">29</div>
+        <div class="bg-info m-2">30</div>
     </div>
 
-    <?php include("./views/common/load_js_scripts.php");?>
+    <?php include($pathToRootFolder."views/common/load_js_scripts.php");?>
 </body>
 
 </html>